% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{package}
\name{Rpickands-package}
\alias{Rpickands-package}
\title{Simulations estimating Pickands constant.}
\description{
Simulations estimating Pickands constant.
}
\details{
To be written.
}
\author{
Kamil Marcin Kosinski [aut]\cr
Marcin Piotr Kosinski [aut, cre]
}
\seealso{
Other Rpickands: \code{\link{Take_expected_value}}
}

