% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{Rpickands-package}
\alias{Rpickands-package}
\title{Simulations estimating Pickands constant.}
\description{
Simulations estimating Pickands constant.
}
\details{
To be written.
}
\author{
Kamil Marcin Kosinski [aut] \email{kamil.kosinski@gmail.com} \cr
Marcin Piotr Kosinski [aut, cre] \email{m.p.kosinski@gmail.com}
}
\seealso{
Other Rpickands: \code{\link{Take_expected_value}}
}

